# api/openapiv2/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "kiae",
    "version": "v0.0.1",
    "contact": {
      "name": "kiae",
      "url": "https://github.com/kiaedev/kiae"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/apps": {
      "get": {
        "operationId": "AppService_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pid",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AppService"
        ]
      },
      "post": {
        "operationId": "AppService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appApplication"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appApplication"
            }
          }
        ],
        "tags": [
          "AppService"
        ]
      }
    },
    "/api/v1/apps/{id}": {
      "delete": {
        "summary": "rpc Start (AppStatusRequest) returns (AppStatusReply) {\n       option (google.api.http) = {\n           patch: \"/api/v1/apps/{id}/status\"\n           body: \"*\"\n       };\n   }\n   rpc Stop (AppStatusRequest) returns (AppStatusReply) {\n       option (google.api.http) = {\n           patch: \"/api/v1/apps/{id}/status\"\n           body: \"*\"\n       };\n   }",
        "operationId": "AppService_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appAppOpReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "op",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "OPERATION_UNSPECIFIED",
              "OPERATION_INSTALL",
              "OPERATION_UNINSTALL",
              "OPERATION_UPGRADE"
            ],
            "default": "OPERATION_UNSPECIFIED"
          }
        ],
        "tags": [
          "AppService"
        ]
      },
      "put": {
        "operationId": "AppService_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appApplication"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "pid": {
                  "type": "string"
                },
                "env": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "ports": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/projectPort"
                  }
                },
                "status": {
                  "$ref": "#/definitions/appStatus"
                },
                "size": {
                  "$ref": "#/definitions/appSize"
                },
                "configs": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/projectConfiguration"
                  }
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "title": "string depends = 8;"
                },
                "createdBy": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedBy": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        ],
        "tags": [
          "AppService"
        ]
      },
      "patch": {
        "operationId": "AppService_Update2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appApplication"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "pid": {
                  "type": "string"
                },
                "env": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "image": {
                  "type": "string"
                },
                "ports": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/projectPort"
                  }
                },
                "status": {
                  "$ref": "#/definitions/appStatus"
                },
                "size": {
                  "$ref": "#/definitions/appSize"
                },
                "configs": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/projectConfiguration"
                  }
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "title": "string depends = 8;"
                },
                "createdBy": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedBy": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        ],
        "tags": [
          "AppService"
        ]
      }
    },
...
  },
  "definitions": {
    "appAppOpReply": {
      "type": "object"
    },
...
    "appApplication": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "pid": {
          "type": "string"
        },
        "env": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/projectPort"
          }
        },
        "status": {
          "$ref": "#/definitions/appStatus"
        },
        "size": {
          "$ref": "#/definitions/appSize"
        },
        "configs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/projectConfiguration"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "title": "string depends = 8;"
        },
        "createdBy": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedBy": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "appListResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/appApplication"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "appSize": {
      "type": "string",
      "enum": [
        "SIZE_NANO",
        "SIZE_MIRCO",
        "SIZE_MINI",
        "SIZE_SMALL",
        "SIZE_MEDIUM",
        "SIZE_LARGE",
        "SIZE_XLARGE",
        "SIZE_XXLARGE"
      ],
      "default": "SIZE_NANO"
    },
    "appStatus": {
      "type": "string",
      "enum": [
        "STATUS_PENDING",
        "STATUS_RUNNING",
        "STATUS_STOPPED"
      ],
      "default": "STATUS_PENDING"
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "projectConfiguration": {
      "type": "object",
      "properties": {
        "filename": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "mountPath": {
          "type": "string"
        }
      }
    },
...
    "projectPort": {
      "type": "object",
      "properties": {
        "port": {
          "type": "integer",
          "format": "int64"
        },
        "protocol": {
          "type": "string"
        }
      }
    },
...
  }
}